refactor: Moved models_enums.py to src/app/models/dqm1 directory
